Server code cleaned But I have lost track of how many things I was doing
