+++ src/client/components/pages/RegisterPage.js
@@ -10,7 +10,7 @@ import React, { Component} from 'react';
 import { connect } from 'react-redux';
 import Form_Register_Student from '../Form_Register_Student.js';
 import { sendingRequest, register } from '../../actions/AppActions';
-import LoadingIndicator from '../LoadingIndicator.js';
+import Loadingindicator from '../Loadingindicator.js';
 
 class RegisterPage extends Component {
     render() {
